+ just optional explanations of pre update(that crucial walala...)
+) I mean, Also Crucial Too.... update means there are alarm function in pointMarks.
Now, they have alarm function.
When the snake and Point are at the same line, now the Point glow brightly in green color. (tip: red and green is perfectly opposite color)
So player can immediately notice when in the situation like "Can i eat the point if i directly rush in this line? or maybe move 1 pixel left and then rush?"
Actually this func is crucial too. Now all crucial functions are ready(maybe), so i call this V2.0