Last update) Bug fixed and Roll back
Sadly to say, this update is the last- (maybe)
Beacuse my eyes are very very at low condition and the reason is this devil snake.

Game graphic is (Actually) not good at eyes, but if you enjoy this game short time like 2~3 minitues (or more less) it doesn't care

But i spend many more time for watching and monitoring this VHS CRT NEON CYBERPUNK theme screen.

not Good, anyway. i love this game but i hate this game. AMBIVALENT

To get back to the story, 
the pre update(view angle, cityline etc.) was not my type.(bad at Dry eyes) so there is roll-back here.
And the sound-audio issue when die and retry(that exists at V2.0) has fixed.

Good night.